% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catego.R
\name{catego}
\alias{catego}
\alias{catego.numeric}
\alias{catego.formula}
\title{Compute pairwise tests and categorize groups from the results of the pairwise function}
\usage{
catego(x, ...)

\method{catego}{numeric}(
  x,
  g,
  test_fnct = "t.test",
  p.adjust.method = p.adjust.methods,
  conf.level = 0.95,
  ...
)

\method{catego}{formula}(
  formula,
  test_fnct = "t.test",
  p.adjust.method = p.adjust.methods,
  conf.level = 0.95,
  data,
  ...
)
}
\arguments{
\item{x}{the values : a vector of values, off the samples to be compared, of a formula.}

\item{...}{other parameters to give to \code{pairwise.table()}.}

\item{g}{the groups : the vector of factors, same length than parameter \code{x}}

\item{test_fnct}{the fonction used to calculate p-value. Actually
't.test', 'wilcox.test', 'yuen','var.test' are supported.}

\item{p.adjust.method}{the method to correct the p-value.
See \code{p.ajust.methods} to see the whole list.}

\item{formula}{the formula giving  \code{x~g} where \code{x} are the quantitatives measures
and \code{g} is a factor of the same length.}

\item{data}{the dataframe where \code{x} and \code{g} are
columns names.}
}
\value{
A list  :
\itemize{
\item a dataframe $cat with a named category for each group,
\item an object of class \code{pairwise.htest}.
If a group are between two category, and are
name with the two letters.
}
}
\description{
Compute pairwise tests and categorize groups from the results of the pairwise function
}
\examples{
#catego(x=iris$Sepal.Length, g=iris$Species, p.adjust.method='bonferroni')
}
\author{
Julien Bousquet (2021)
}
