% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.compare.R
\name{device.print}
\alias{device.print}
\title{Plot data.frame and titles}
\usage{
device.print(x, y, info = NULL, title = NULL)
}
\arguments{
\item{x}{abscisse position of print}

\item{y}{ordinate position of print}

\item{info}{the information to print, a data.frame or a character string}

\item{title}{a character string title printed in bald font. If only title is given, print in bald and italic}
}
\description{
To help print graphics in m.test()
}
\author{
Julien Bousquet (2021)
}
