% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon.cut.test.R
\name{ind.median}
\alias{ind.median}
\title{Indexes around median.}
\usage{
ind.median(group, prop)
}
\arguments{
\item{group}{a vector with a calculable median}

\item{prop}{a proportion between 0 and 1 of values to find around the median. The median is always returned among the indexes, and the proportion is symetricaly define around it.}
}
\value{
The function return the vector of indexes of prop values around the median of the group
}
\description{
Compute index of a proportion of values of group around the median
}
\details{
ind.median() is used in wilcoxon.cut.test() when prop>0.
The value of the median is always excluded.
}
\author{
Julien Bousquet (2021)
}
