% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.compare.R
\name{lm.compare}
\alias{lm.compare}
\title{Compare two linear models}
\usage{
lm.compare(
  reg1,
  reg2,
  nested = c(TRUE, FALSE),
  code = TRUE,
  verbose = TRUE,
  conf.level = 0.95
)
}
\arguments{
\item{reg1}{,reg2 two  models to compare.}
}
\description{
Execute the tests of  \code{lmtest} package to compare
two linear model, fitted with \code{lm()}.
}
