% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.test.R
\name{lm.test}
\alias{lm.test}
\title{Execute the tests of  \code{lmtest} package to validate
a fit of linear model with \code{lm()}.}
\usage{
lm.test(reg, order.by = NULL, code = TRUE, verbose = TRUE, conf.level = 0.95)
}
\arguments{
\item{reg}{the result of a linear model, obtained with \code{lm()}.}

\item{order.by}{a vector z. The
observations in the model are ordered by the size of z.
If set to NULL (the default), the observations are
assumed to be ordered (e.g., a time series).}

\item{code}{boolean. If TRUE, write the code to replay, in the response.}

\item{verbose}{boolean. If TRUE, write a short remainder comments about
the interpretation of each test.}
}
\description{
Execute the tests of  \code{lmtest} package to validate
a fit of linear model with \code{lm()}.
}
