% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.compare.R
\name{mean.compare}
\alias{mean.compare}
\title{Automatic mean/median comparison}
\usage{
\method{mean}{compare}(
  x,
  g,
  pval = 0.05,
  verbose = FALSE,
  return = TRUE,
  paired = FALSE,
  pval_ks = 0.01,
  maxcat = 15,
  plot = TRUE,
  silent = TRUE,
  boot = FALSE,
  iter = 100,
  conf = 0.95,
  code = FALSE,
  debug. = FALSE
)
}
\arguments{
\item{x}{the values : a vector of values, with the samples to be compared.}

\item{g}{groups : a vector of less (than maxcat) factors}

\item{pval}{the usual level on confidence, 0.05 by default.}

\item{verbose}{\code{FALSE} by defaul. Increase information.}

\item{return}{to be completed}

\item{paired}{FALSE by default, but can be passed to TRUE if each value of x in categories are paired and in the same order.}

\item{pval_ks}{The p-value for repartitions tests like Kolmogorov-Smirnov or Shapiro :  0.01 by default.}

\item{maxcat}{The maximum number of categories : 50 by default.}

\item{plot}{\code{TRUE} to plot the graph with all the results.}

\item{silent}{\code{TRUE} to keep the function silent : no warning}

\item{iter}{100 by default, number of iterations in the bootstrap.}

\item{conf}{Confidence of the bootstrap, 0.95 by default.}

\item{code}{if \code{TRUE}, return the code that makes the different tests. \code{FALSE} by default}

\item{bootstrap}{\code{FALSE} by default, dont do bootstrap. \code{TRUE} to make \code{iter} sampling}
}
\description{
\code{mean.compare()} do all the possible tests of comparison of mean, even not mathematically corrects and even
if hypothesis are wrong. Execute tests to verify hypothesis. The results are summarised in a graph. The user need
to choose the good path, which is be done automatically by \code{mean.compare()}. By default, find the path the more to the left.
}
\examples{
#mean.compare(x=iris$Petal.Length, g=iris$Species)
}
\author{
Julien Bousquet (2021)
}
