% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catego.R
\name{pairwise}
\alias{pairwise}
\alias{pairwise.numeric}
\alias{pairwise.formula}
\title{Pairwise comparisons tests}
\usage{
pairwise(x, test_fnct = "t.test", p.adjust.method = p.adjust.methods, ...)

\method{pairwise}{numeric}(x, g, test_fnct = "t.test", p.adjust.method = p.adjust.methods, ...)

\method{pairwise}{formula}(
  formula,
  test_fnct = "t.test",
  p.adjust.method = p.adjust.methods,
  data,
  ...
)
}
\arguments{
\item{x}{the values : a vector of values, off the samples to be compared, of a formula.}

\item{test_fnct}{the fonction used to calculate p-value. Actually
't.test', 'wilcox.test', 'yuen','var.test' are supported.}

\item{p.adjust.method}{the method to correct the p-value.
See \code{p.ajust.methods} to see the whole list.}

\item{...}{other parameters to give to \code{pairwise.table()}.}

\item{g}{the groups : the vector of factors, same length than parameter \code{x}}

\item{formula}{the formula giving  \code{x~g} where \code{x} are the quantitatives measures
and \code{g} is a factor of the same length.}

\item{data}{the dataframe where \code{x} and \code{g} are
columns names.}
}
\description{
\code{pairwise()} compute p-value for all the possible
combinaison of groups. The p-values are corrected by
the \code{pairwise.table()} function of package stats.
}
\examples{
# Classical Student test :
#pairwise(x=iris$Sepal.Length, g=iris$Species, test_fnct='t.test')

# Robust Yuen's test of WRS2 package :
#pairwise(x=iris$Sepal.Length,g=iris$Species, test_fnct='yuen')

# Also compare variance :
#pairwise(x=iris$Sepal.Length,g=iris$Species, test_fnct='var.test')

}
\author{
Julien Bousquet (2021)
}
