% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{star.lm}
\alias{star.lm}
\title{Get stars for a linear model}
\usage{
\method{star}{lm}(p.val, ...)
}
\arguments{
\item{p.val}{a linear model}
}
\description{
Get stars for a linear model
}
\details{
Use the function glance from package broom to get the p.value of
the linear model, and then convert it into stars with function star().
}
