% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{star}
\alias{star}
\title{A function that transforms pvalues in stars ***}
\usage{
star(p.val, probs = c(0.001, 0.01, 0.05, 0.1))
}
\arguments{
\item{p.val}{numeric, the value of the pvalue.}

\item{probs}{a vector of the four probabilities of changes in increasing order.}
}
\value{
A character string with stars, dot or blank space.
}
\description{
A function that transforms pvalues in stars ***
}
\examples{
star(0.015)
#for a vector of p-values
pvals <- c(0.01, 0.05, 0.015, 0.5)
sapply(pvals, star)
}
\author{
Julien Bousquet (2021)
}
