% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testsuite.R
\docType{package}
\name{testsuite}
\alias{testsuite}
\title{testsuite : A package for computing tests in 'the' right way}
\description{
The testsuite package (try to) tests the validaty of hypothesis
(homoscedasticity, normality, or whatever) in the right order,
and execute the tests to compare means, or linear models
}
\section{The fonctions of meantest are}{

mean.compare() executes every test and represent every possible path
pairwise() calculate test between every groups, with a p-value correction
catego() use pairwise to attribute categories to each group.
lm.test() validate linear model
lm.compare() compare nested or non-nested models
}

