% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{warns}
\alias{warns}
\title{Write a warning about the result of a test}
\usage{
warns(name, signe = c("<", ">"), pval, conf.level = 0.95)
}
\arguments{
\item{name}{the name of the test.}

\item{signe}{one character '>' or '<' giving the expected}

\item{pval}{a numeric, between 0 and 1 : the resulting p-value of a test.}

\item{conf.level}{level of confidence for decision.
position of \code{pval}.
expected way for the p-value to be}
}
\description{
Use color to emphasis the result of the test.
}
\author{
Julien Bousquet (2023)
}
