% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanFactor}
\alias{cleanFactor}
\title{Clean factors}
\usage{
cleanFactor(x, g)
}
\arguments{
\item{x}{a vector of values, corresponding to g.}

\item{g}{a vector of categories, which must be checked.}
}
\value{
Give comprehensive error messages.
Return a list with x and g, cleaned if possible.
}
\description{
\code{cleanFactor()} check that the factor given in argument
allows the application of statistical test.
}
\examples{
cleanFactor( 1:6, factor(rep(c('a', 'b', NA), times=2) ))
}
\author{
Julien Bousquet (2023)
}
